prodops-ai-backend/requirements.txt
Flask==2.3.2
spacy==3.7.6
openai==0.28
python-dotenv==1.0.0